+++ app/src/main/java/com/theandroiddev/mywins/local/database/AppDatabase.java
@@ -5,8 +5,6 @@ import android.arch.persistence.room.RoomDatabase;
 
 import com.theandroiddev.mywins.local.dao.SuccessDao;
 import com.theandroiddev.mywins.local.dao.SuccessImageDao;
-import com.theandroiddev.mywins.data.model.SuccessEntity;
-import com.theandroiddev.mywins.data.model.SuccessImageEntity;
 import com.theandroiddev.mywins.local.model.LocalSuccess;
 import com.theandroiddev.mywins.local.model.LocalSuccessImage;
