+++ app/src/main/java/com/theandroiddev/mywins/local/model/LocalSuccess.kt
@@ -3,8 +3,7 @@ package com.theandroiddev.mywins.local.model
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
-import com.theandroiddev.mywins.data.model.SuccessEntity
-import com.theandroiddev.mywins.domain.service.successes.SuccessesServiceModel
+import com.theandroiddev.mywins.data.successes.model.SuccessEntity
 import com.theandroiddev.mywins.utils.Constants
 import java.io.Serializable
