+++ app/src/main/java/com/theandroiddev/mywins/local/model/LocalSuccessImage.kt
@@ -3,8 +3,7 @@ package com.theandroiddev.mywins.local.model
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
-import com.theandroiddev.mywins.data.model.SuccessImageEntity
-import com.theandroiddev.mywins.domain.service.success_images.SuccessImagesServiceModel
+import com.theandroiddev.mywins.data.successes.model.SuccessImageEntity
 
 @Entity
 data class LocalSuccessImage(
